Test fun for down case the text file

+++ src/main/kotlin/prod/GiveMeaWord/Repostitory/WordsRepository.kt
@@ -5,13 +5,13 @@ import org.springframework.stereotype.Repository
 import prod.GiveMeaWord.ModelBase.Word
 
 @Repository
-interface WordsRepository: CrudRepository<Word, Long> {
+interface WordsRepository: CrudRepository<Word, Int> {
 
     fun save(word: Word)
     fun findByLength(length:Int):Iterable<Word>
     fun existsWordByWord(word:String):Boolean
-    fun getAllByIdAfter(id: Long):Iterable<Word>
-    fun findOneById(id: Long):Word
+    fun getAllByIdAfter(id: Int):Iterable<Word>
+    fun findOneById(id: Int):Word
     override fun count():Long
     override fun deleteAll()
 }
